solve clang-tidy basic warning

+++ include/condy/work_type.hpp
@@ -6,7 +6,7 @@
 
 namespace condy {
 
-enum class WorkType : intptr_t {
+enum class WorkType : uint8_t {
     Common,
     Ignore,
     Notify,
@@ -19,6 +19,7 @@ inline std::pair<void *, WorkType> decode_work(void *ptr) {
     intptr_t mask = (1 << 3) - 1;
     intptr_t addr = reinterpret_cast<intptr_t>(ptr);
     WorkType type = static_cast<WorkType>(addr & mask);
+    // NOLINTNEXTLINE(performance-no-int-to-ptr)
     void *actual_ptr = reinterpret_cast<void *>(addr & (~mask));
     return std::make_pair(actual_ptr, type);
 }
@@ -28,6 +29,7 @@ inline void *encode_work(void *ptr, WorkType type) {
     // Ensure align of 8
     assert(addr % 8 == 0);
     addr |= static_cast<intptr_t>(type);
+    // NOLINTNEXTLINE(performance-no-int-to-ptr)
     return reinterpret_cast<void *>(addr);
 }
